Move `OsuColors` package
